move participation permission check into participation service
