+++ src/main/java/de/tum/in/www1/artemis/service/ParticipationService.java
@@ -12,6 +12,8 @@ import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
+import javax.annotation.Nullable;
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Value;
@@ -77,11 +79,13 @@ public class ParticipationService {
 
     private final ModelAssessmentConflictService conflictService;
 
+    private final AuthorizationCheckService authCheckService;
+
     public ParticipationService(ParticipationRepository participationRepository, ExerciseRepository exerciseRepository, ResultRepository resultRepository,
             SubmissionRepository submissionRepository, ComplaintResponseRepository complaintResponseRepository, ComplaintRepository complaintRepository,
             QuizSubmissionService quizSubmissionService, ProgrammingExerciseRepository programmingExerciseRepository, UserService userService, Optional<GitService> gitService,
             Optional<ContinuousIntegrationService> continuousIntegrationService, Optional<VersionControlService> versionControlService,
-            SimpMessageSendingOperations messagingTemplate, ModelAssessmentConflictService conflictService) {
+            SimpMessageSendingOperations messagingTemplate, ModelAssessmentConflictService conflictService, AuthorizationCheckService authCheckService) {
         this.participationRepository = participationRepository;
         this.exerciseRepository = exerciseRepository;
         this.resultRepository = resultRepository;
@@ -96,6 +100,7 @@ public class ParticipationService {
         this.versionControlService = versionControlService;
         this.messagingTemplate = messagingTemplate;
         this.conflictService = conflictService;
+        this.authCheckService = authCheckService;
     }
 
     /**
@@ -699,5 +704,44 @@ public class ParticipationService {
 
     public Participation findOneWithEagerCourse(Long participationId) {
         return participationRepository.findOneByIdWithEagerExerciseAndEagerCourse(participationId);
+    }
+
+    /**
+     * Check if a participation can be accessed with the current user.
+     *
+     * @param participation
+     * @return
+     */
+    @Nullable
+    public boolean canAccessParticipation(Participation participation) {
+        if (!userHasPermissions(participation))
+            return false;
+
+        if (!Optional.ofNullable(participation).isPresent()) {
+            return false;
+        }
+        return true;
+    }
+
+    /**
+     * Check if a user has permissions to to access a certain participation. This includes not only the owner of the participation but also the TAs and instructors of the course.
+     *
+     * @param participation
+     * @return
+     */
+    private boolean userHasPermissions(Participation participation) {
+        if (!authCheckService.isOwnerOfParticipation(participation)) {
+            // if the user is not the owner of the participation, the user can only see it in case he is
+            // a teaching assistant or an instructor of the course, or in case he is admin
+            User user = userService.getUserWithGroupsAndAuthorities();
+            // TODO: temporary workaround for problems with the relationship between exercise and participations / templateParticipation / solutionParticipation
+            if (participation.getExercise() == null) {
+                // this can only happen if we have a template or solution participation. Then the call can only be invoked by an instructor / teaching assistant
+                return true;
+            }
+            Course course = participation.getExercise().getCourse();
+            return authCheckService.isTeachingAssistantInCourse(course, user) || authCheckService.isInstructorInCourse(course, user) || authCheckService.isAdmin();
+        }
+        return true;
     }
 }
