+++ src/main/java/de/tum/in/www1/artemis/web/rest/ResultResource.java
@@ -246,7 +246,7 @@ public class ResultResource {
         List<Result> results = new ArrayList<>();
         Participation participation = participationService.findOne(participationId);
 
-        //TODO: temporary workaround for problems with the relationship between exercise and participations / templateParticipation / solutionParticipation
+        // TODO: temporary workaround for problems with the relationship between exercise and participations / templateParticipation / solutionParticipation
         if (participation.getExercise() == null) {
             Exercise exercise = exerciseService.findOne(exerciseId);
             participation.setExercise(exercise);
@@ -409,10 +409,9 @@ public class ResultResource {
     @PreAuthorize("hasAnyRole('TA', 'INSTRUCTOR', 'ADMIN')")
     public ResponseEntity<Result> getLatestResultWithFeedbacks(@PathVariable Long participationId) {
         log.debug("REST request to get latest result for participation : {}", participationId);
-        User user = userService.getUserWithGroupsAndAuthorities();
         Participation participation = participationService.findOne(participationId);
 
-        if (!authCheckService.isAtLeastTeachingAssistantInCourse(participation.getExercise().getCourse(), user)) {
+        if (!participationService.canAccessParticipation(participation)) {
             return forbidden();
         }
 
@@ -437,19 +436,9 @@ public class ResultResource {
             return notFound();
         }
         Participation participation = result.get().getParticipation();
-        Course course = participation.getExercise().getCourse();
 
-        if (participation.getStudent() == null) {
-            // If the student is null, then we participation is a template/solution participation -> check for instructor role
-            if (!authCheckService.isAtLeastInstructorForCourse(participation.getExercise().getCourse(), null)) {
-                return forbidden();
-            }
-        }
-        else {
-            if (!authCheckService.isOwnerOfParticipation(participation)) {
-                if (!userHasPermissions(course))
-                    return forbidden();
-            }
+        if (!participationService.canAccessParticipation(participation)) {
+            return forbidden();
         }
 
         try {
